+++ .gitignore
@@ -1,3 +1,2 @@
 # This ignores all initialization files. DO NOT CHANGE THIS FILE!
 db-storage
-node_modules
